qfpython/requirements.txt
aliyun-python-sdk-core-v3==2.13.10
certifi==2019.9.11
chardet==3.0.4
Django==2.0
django-debug-toolbar==2.1
django-haystack==2.8.1
django-redis==4.10.0
django-restframework==0.0.1
django-shortuuidfield==0.1.3
djangorestframework==3.10.3
idna==2.8
jieba==0.39
jmespath==0.9.4
mysqlclient==1.4.4
Pillow==6.1.0
python-memcached==1.59
pytz==2019.2
qiniu==7.2.6
redis==3.3.11
requests==2.22.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
Whoosh==2.7.4
